Factor out the progress display utilities
