+++ lib/convert.js
@@ -2,7 +2,7 @@
 const path = require("path");
 const fs = require("fs").promises;
 const workerpool = require("workerpool");
-const cliProgress = require("cli-progress");
+const progressUtils = require("./progress-utils.js");
 
 module.exports = async (
   cachePath,
@@ -20,14 +20,7 @@ module.exports = async (
   const flattenedChapters = chapterData.flatMap(arc => arc.chapters);
 
   console.log("Converting raw downloaded HTML to EPUB chapters");
-  const progress = new cliProgress.SingleBar({
-    stopOnComplete: true,
-    clearOnComplete: true,
-    format: " {bar} {percentage}% | {time} | {value}/{total}"
-  }, cliProgress.Presets.shades_classic);
-
-  const start = performance.now();
-  progress.start(flattenedChapters.length, 0, { time: "     " });
+  const progress = progressUtils.start(flattenedChapters.length);
 
   const poolOptions = {};
   if (concurrentJobs !== undefined) {
@@ -42,8 +35,7 @@ module.exports = async (
 
     warnings.push(...await pool.exec("convertChapter", [chapter, bookData.title, inputPath, outputPath]));
 
-    const seconds = String(Math.round((performance.now() - start) / 1000)).padStart(3);
-    progress.increment({ time: `${seconds} s` });
+    progressUtils.increment(progress);
   }));
 
   pool.terminate();
@@ -52,7 +44,7 @@ module.exports = async (
     console.warn(warning);
   }
 
-  console.log(`All chapters converted in ${Math.round((performance.now() - start) / 100) / 10} seconds`);
+  console.log(`All chapters converted in ${progressUtils.getTotalSeconds(progress)} seconds`);
 };
 
 function getChapterData(arcs, manifest, chapterTitleStyle) {
