+++ lib/download.js
@@ -2,7 +2,7 @@
 const path = require("path");
 const fs = require("fs").promises;
 const { JSDOM } = require("jsdom");
-const cliProgress = require("cli-progress");
+const progressUtils = require("./progress-utils.js");
 
 const FILENAME_PREFIX = "chapter";
 
@@ -58,13 +58,7 @@ async function downloadAllChapters(manifest, bookData, cachePath, manifestPath) 
 
   console.log(`Downloading ${toDownloadManifestIndicesByURL.size} remaining HTML pages`);
 
-  const progress = new cliProgress.SingleBar({
-    stopOnComplete: true,
-    clearOnComplete: true,
-    format: " {bar} {percentage}% | {time} | {value}/{total}"
-  }, cliProgress.Presets.shades_classic);
-  const start = performance.now();
-  progress.start(toDownloadManifestIndicesByURL.size, 0, { time: "     " });
+  const progress = progressUtils.start(toDownloadManifestIndicesByURL.size);
 
   const promises = [];
   let manifestUpdatePromise = null;
@@ -93,13 +87,12 @@ async function downloadAllChapters(manifest, bookData, cachePath, manifestPath) 
         });
       }
 
-      const seconds = String(Math.round((performance.now() - start) / 1000)).padStart(3);
-      progress.increment({ time: `${seconds} s` });
+      progressUtils.increment(progress);
     })());
   }
 
   await Promise.all(promises);
-  console.log(`All HTML pages downloaded in ${Math.round((performance.now() - start) / 100) / 10} seconds`);
+  console.log(`All HTML pages downloaded in ${progressUtils.getTotalSeconds(progress)} seconds`);
 }
 
 function getChapterTitle(rawChapterDoc) {
